requirements.txt
asyncio
websockets
uuid
mcp
python-dotenv
fastapi
uvicorn
pydantic